Simplify BossDoor Interaction
changed how hero interacts with bossDoor.
player must now press A button while hero is facing bossDoor to open it.
this occurs via an InteractionRec collision, removing bossDoor from actor collision.
this was done because the bossDoor is blocking, and we want to remove all
blocking objects from the Interact(Actor, Obj) method in prep for an optimization.