Improve Picking Up Room Objects in Editor
shouldn't be able to grab particles or pickups in editor.
there was an issue where i couldn't pickup an obj because a particle overlapped it.
it was annoying, and i don't see a reason why i would ever need to pickup a particle.
updated ObjectToolsWidget.GrabRoomObject() to ignore particles & pickups in Entity Pool.
playtested, working as expected.