Optimize Animated Pit Bubbles
renamed ParticleBubble to ParticlePitAnimation.
because in the future this animation may not necessarily be a bubble.
we may change the animation in the dungeon sheet to be something else.
like lava bubbling, or waves, or whatever.
so the name needs to be generic.
increased lifetime of ParticlePitAnimation to 255, so they last longer.
decreased how often these particles are created in Pits.
this has the net effect of using less particles to display about the same amount.
playtested, various polishing of values, working/looking great.