Polish Pit Interaction with Hero
when hero collides with a pit, the hero's shadow is visible.
this looks visually incorrect, as the shadow should be under hero at all times.
so we simply hide the hero's shadow if he collides with a pit.
when the pit collision is over, we unhide hero's shadow.

updated Interact() -> Pit vs Hero to hide/unhide hero's shadow.
playtested, working as expected.