DialogScreen++
removed boolean PauseDungeonScreen from DialogScreen.
this boolean controlled if the Dungeon/Level screen received updates from DialogScreen.
this was being done so that the hero could finish his animation.
for example, when the hero open the map chest, he got the map.
he holds it above it his head (reward anim) and dialog screen opens.
however, if we pause the level screen, then the hero doesn't have enough time
to transition into his reward animation, so it looks wrong.

updated DialogScreen to set hero into his current animation when the screen opens.
this solves the problem of the hero's sprite being displayed incorrectly.
plus, we no longer have to update the level screen (unpause game).
if we continued to run the game while the dialog screen was open,
it prevented player from passing input to hero, while enemies were still moving around.
this led to a state where the enemies could attack and kill the hero,
and the player couldn't do anything about it, which sucks.
so that had to be fixed.

playtested, working as expected.