Implement Self-Rez/Heal upon Death
playtesting, a combo potion was used upon death - which was awesome!
this was a bug, it was checking the fairy bottles old value, but it felt right.
we should use fairys, combos, and health potions upon death, if hero has them.
this prevents the problem of having to pause and use a health potion while in battle.
i've died in LttP when I had a health potion, and it was frustrating.

created Functions_Item.CheckBottleUponDeath().
this checks all bottles for health, combo, and fairy and uses them.
this is called from Functions_Actor.Update() -> hero death state.
playtested, major bug found.

Big Old Bug
during normal RELEASE gameplay, dying in a dungeon leads to summary screen.
this leads to overworld screen, but we're hearing actor sounds during this screen.
cause: the overworld screen was calling Pool.Update(), which was updating everything.
the overworld screen only needs to update the entity pool.
split Pool.Update() into UpdateActors(), UpdateEntities(), and UpdateRoomObjects().
this allows OverworldScreen to only call UdpateEntities().
playtested, big old bug fixed.