Various Improvements
reduced fairy drop rate to 1%, down from 5%.

bug: we can move hero when he is in a self-rez reward state.
this happens when hero uses a bottle upon death.
created Functions_Actor.SetRewardState(Actor).
this standardizes how an actor is set into the reward state.
this ensures that the actor is statelocked.
the lack of statelock during selfrez reward was causing the moving bug.

removed door collapse debris.
it didn't look that good anyway.
plus, door debris was located on the main sheet and took up 4 frames.
it should of been located on the dungeon sheet (for themeing).
but i dont want to waste 4 frames on a door collapse animation.
so we dont have additional fx for door collapse, just the attention particle.
which is fine.

trim all sound files as much as possible (no lead/tail).
updated dX and UWP projects with trimmed soundFX.
updated equipment vendor to only display working equipment.