Fix Room Designs
fixed two square rooms that weren't playing as I'd like them to.
the main issue was traversing them as a blob, where blob hero could get stuck.
the other issue was simply that too much bouncing was happening in the room.
so I removed some bumpers and replaced them with skull pots.