Display SkillRating's Change on SummaryScreen
display the amount the player's rating has changed, on summary screen.
if the amount is positive, a "+" should be added before the value.
the value should be trimmed to 2 digits.

created SummaryScreenV2 mockup.
updated SS to calculate newSR & oldSR, then display difference.
fixed spacing of "to continue" text to center to screen.
playtested, working as expected.