Handle Dungeon Exits with a Summary Screen
summary screens should be displayed when hero beats, exits, or dies in a dungeon.
for a level like the shop, the summary screen should never be displayed.

updated Functions_Interaction.InteractActor() doors section.
added Level.type check for Castle, which changed the exit action to Summary.
otherwise, the exit action defaults to Overworld.
also fixed an invisible bug where the hero..
was not being stopped upon collision with any exit door.
playtested, working but with some bugs and missing states.

issue: when the hero exits a dungeon, the "you died" text appears.
this is incorrect, as the hero did not die. use "defeat" text instead.
this requires an addition to the text sprites sheet.

bug: there is no distinction in code between hero exiting from or dying in a dungeon.
both branches are handled the same way, causing the level screen to fade out very slowly.
if the hero is exiting a level, then the fade should always be quick.
we need to be able to tell the difference between an exit, and a hero win/loss.
added ExitDungeon value to ExitAction enum.
updated LevelScreen.Update() to handle new ExitAction branch for ExitDungeon.
this code section was then refactored and simplified to only check for Summary state.
added CloseLevel(ExitAction) to Functions_Level.
simplified various code sections to use CloseLevel(ExitAction).
in all code sections this reduced and simplified code.
playtested, working as expected.