Fix ExitObj Collision Bug
if hero dashes right before a collision with an exit obj,
the hero can dash thru the door and appear beyond the level.
this happens for just a moment, and the level exits properly.
but it doesn't look correct.

upon collision with an exit obj, the hero should stop moving.
updated Functions_Interaction.
playtested, working as expected.