Fix Improper Display of Elapsed Game Time
the time is not being displayed properly on the game display.
the seconds are being added to, and they exceed 60.
the issue stems from how we add a timeSpan to our PlayerData.
we are not rounding the secs & mins around a radix of 60.
so those values can exceed the bounds of 60.

updated SummaryScreen to round the secs and mins values.
playtested, working as expected.