Fix Sword Projectile on Conveyor Belt Bug
projectiles like swords dont move on conveyor belts.
so if the hero attacks with a sword facing south on a conveyor belt,
the hero will be pushed into his own sword, dealing him damage.
obviously, this behavior is wrong and bad.

when an actor attacks, uses an item, or gets rewarded,
halt their movement if colliding with a conveyor belt.
else, push the actor as normal.

updated Functions_Interaction with code to check for actor state.
based on actor state, actors are pushed or not pushed on treadmills.
playtested, working as expected.