Implement Torch Room Puzzle
added Torches value to PuzzleType enum.
added code for torch counting to CheckForPuzzles().
rooms with more than 3 torches are set to Torch PuzzleType.
doors are converted to trap doors as well.

now when hero lights a torch, the game checks to see if,
lighting that torch solved the room puzzle, and opens the doors if so.
added Function_RoomObject.CountTorches().

playtested that 4 torches creates and opens a puzzle trap room.
changed how trap door are sorted to appear more natural.