Hero should drop any Pot he is carrying upon a Pit collision
moved code to drop the hero's carryingObj into method DropCarryingObj().
simplified Hero.HandleState() to call Hero.DropCarryingObj().
this was done because we need to be able to call DropCarryingObj(),
when hero falls into a pit.
updated InteractActor() with code to handle dropping of hero's carrying obj.
also simplified the PitAnimated regions a bit.
playtested, working as expected.
the hero does drop the obj, and does not respawn with it.