Improve Actor / Obj falling into Pits
changed the percentage scale value required to trigger end case for actor/pit interaction.
in english, this means the actor doesn't scale down to 0% when colliding with a pit.
it visually looked strange to watch the actor disappear into an abyss, when clearly we
can see the bubbles on the surface of the pit liquid.
instead, the actor / object triggers when it reaches 80% scale.
playtested, looks / feels much better.