Fix Throwing Pots North Bug
Bug: throwing a pot north creates a shadow projectile that collides with hero.
since all projectiles by default 'hit' hero, this causes the hero to be set into
a hit state, which is buggy and feels cheap and bad.

fix: any shadow projectile collision with actor should be ignored.
this is identical to debris rock projectile collisions, which are ignored.
playtested, working great.