Refactoring Collisions / Interactions - Part 1
i'm not going to even attempt to document all the changes made,
instead i'll just explain from a high level what needed to be done.

the interaction system was tied to the collision system.
when a collision happened, an interaction happened.
with a per axis collision checking system, this means
that an obj could collide with another obj up to 3 times
in a single frame, which would call Interact() 3 times too.
obvs, this behavior is unwanted, and so objs were only
using a single axis collision checking system, which meant
objs couldn't slide against other objs, but atleast their
interactions were perdictable.

but now, we need sliding for roomObjs so conveyor belts work.
upon working with the collision system i realized the interaction
system needed to be decoupled, and therefore a refactor was necessary.
so, how do we fix collisions and decouple interactions?

Push 1 : Seperate blocking collisions and interactions
this works for entites, roomObjs, and actors:
a first pass of blocking checks to resolve the legal movement.
then a second pass of interaction checks to resolve all interactions.
these interactions are non-blocking or blocking,
it doesn't matter once we resolve the legal movement.
	for example, in case hero was standing over a block that just popped up,
	here we would handle that overlap/collision,
	and set the block back to a down state

what is the current process?
	Pool.Update() -> UpdateActors(), UpdateEntites(), UpdateRoomObjs()
	UpdateActors() -> ProjectMovement(), CheckCollisions(Actor), Align()
	UpdateObj() -> SetMoving, ProjectMovement(), CheckCollisions(Obj), Align()

the new process:
	UpdateActors()/UpdateObjs() ->
		check if active, update, animate, scale..
		set moving boolean
		if moving, project movement()
		check collisions(), set move.NewPos (resolve blocking collisions)
		align actor/obj to move.NewPos
		check interaction() overlaps, handle interactions
	this means we can have simply: Update(Actor), Update(Obj)

Rewrote large pieces of:
Functions_Pool, Functions_Collision, and Functions_Interaction,
along with other various classes.