Flamethrower Improvements
updated flamethrower floor objects to be moveable.
this allows belts to push flamethrowers along their paths.
playtested.

flamethrowers now sort over conveyor belts.
this makes it a bit easier to 'read' where the flamethrower is,
when it's being pushed by conveyor belts.
if a fireball collides with an unlit torch, it should be lit.
extracted functionality for lighting torches into RoomObject functions.
called this method when hero interact recs with unlit torch.
AND when fireball collides with unlit torch.
playtested.

added torches and belts DEV room into game as Row_004.