Pit Interactions++
updated Pit interaction with Actors, such that dead
actors will no longer be set into a hit state as they
collide with and fall into a pit. it made no sense for
a dead enemy to suddenly spring back to life as they
fell into a pit. playtested, working as expected.