Split InteractObject() and Refactoring
Why is this being done? To simplify interactions.
currently its a many to many comparison, which is
really, really bad in this case.

split InteractObject() into:
InteractEntities() and InteractRoomObj()
which vastly simplifies the comparison check.

also, found a way to speed up CheckInteractions()
which involves simply checking to see if the
entity or roomObj is active or not.

which led to additional optimizations for the
new collision and interaction systems.
which led to refactoring, etc...