Fix Actors overlapping Hero Bug
if an actor overlaps the hero, then the hero can't move.
however, functionality is different for enemies and pets.

for the pet, we set the pet's collision component's blocking
boolean to false, which removes it from the collision check.
this allows the pet to overlap the hero without blocking him.
this is because we frequently spawn the pet ontop of the hero,
so they will overlap and the game needs to handle this.

for enemies, we treat them as the hero and perform
full actor v actor collision checks, in addition to the
roomObj checks being done, which makes this expensive.
*we could just do enemy v hero collision checks,
which would DRASTICALLY reduce the cost of the enemy
collision checking... hmmm..

split CheckActors() into CheckEnemies() and CheckHero().
updated CheckCollisions() with an additional control boolean.
updated boolean parameter names as well.
lots of playtesting.