Add Projectile Class
created Projectile class, which extends GameObject.
this adds two fields: one for an actor reference,
and a Vector2 for the origin of the projectile.

this allows a projectile to return to the caster,
even if the caster is moving, or the projectile
can return to it's 'birth' position if we need
to fake a boomerang return to a roomObj, for example.