Fix Projectile vs Dead Actor Interactions
simply put, a projectile shouldn't collide or
interact with a dead actor's corpse. this leads
to a state where arrows and fireballs are absorbed
by a dead enemy, which feels cheap and wrong,
since the dead enemy is laying on the ground,
and the flying projectile would pass over them.

updated InteractActor() with check for dead actor.
playtested, working as expected.