Improve Pickups
updated main sheet in Dx and UWP projects,
there was a stray pixel in the Bomb Pickup animation.

updated HandleDeathEvent() for pickups, now the attention
particle is centered to the 8x16 pixel object.
before, there was no offset, and the attention particle was
centering itself to a 16x16 pixel object, which wasn't
the correct size for the 8x16 pixel pickup object.
in this case, the fairy (which is 16x16) can be ignored,
because the fairy never dies, so she would never spawn
an attention particle and never meet the centering routine.
playtested, working as expected.