Fix Fairy Interactions with Blocking Objects
We want to treat any collision with a blocking object as a blocking
interaction for the fairy pickup object.
and upon closer inspection of the interaction system with regards to
blocking roomObj v projectile/pickup, it's actually easier to check
by obj.group, which made it easy to update InteractRoomObj() to prevent
ALL pickups from overlapping / colliding with blocking roomObjs.

however, upon testing this, it's apparent there is a critical flaw:
pickups are not being collision checked (per axis), so any overlap
causes the pickup object to instantly stop moving (instead of sliding),
so it would actually be easier to integrate the fairy as a PROJECTILE
instead of a pickup. which makes sense, because pickups don't move,
thus they aren't per axis collision checked...