Implement Boomerang
this is a projectile that flys in a thrown direction
then upon hitting something or a duration, returns to hero
tracks to hero upon return, and for now, that's it..

added ProjectileBoomerang to ObjType enum.
added GameObjListAnim for Boomerang.
ripped up alot of code regarding Bottles, that I'll have
to inspect later and fix..

updated UseItem() with path for Boomerang.
added def for Boomerang in Obj.SetType().
added path for Boomerang to Projectile.Update().
lots of tuning and playtesting, working really well.
boomerang can be thrown and danced around with.

updated Projectile.Spawn() to only create 1 boomerang at a time.
although, i was playing around with a bunch of boomerangs on screen
and it was hella fun, it was also hella OP. so, awesome but NO.
lots of playtesting.