Boomerang++
moved boomerangInPlay boolean into the Functions_Hero class.
added ResetUponRoomBuild() method to Hero functions.
this is called from Room.BuildRoom(), at the end.
this allows the hero a moment to reset everything to a
useable state, for example: the boomerang could not of
returned to the hero prior to the hero leaving the room,
and without resetting the boomerang's boolean here in the
ResetUponRoomBuild() method, the boomerang would be lost 4ever.
sadFace.jpg.

added sound fx for throwing and collisions for boomerang.
added SoundBoomerangFlying to Dx & UWP projects, built.
updated Assets with boomerang flying instance.
boomerang sounds and feels complete now.. hmmm.