Attack / Use ++
hero should slow down when he's attacking or using an object.
meaning, his friction should increase.

added FrictionSlow field to global World class.
modified Actor.Update() to set an actor's friction to
FrictionSlow if the actor is in a use or attack state.
this causes the hero to slow down a little when he
attacks, which makes the use/attack feel better.
playtested, working as expected.