Improve Bow Animation
when the hero shoots an arrow, a bow particle is created.
in previous implementations, this particle would rotate to
match the facing direction of the hero and the illusion of
the hero using the bow was good. yay.jpg.

however, the new system doesn't allow for the rotation of
particles, PLUS the bow particle should track to the hero
across frames, because the hero can now slide while attacking,
which means the current bow particle stays behind and faces
the wrong direction, which is completely wrong. boo.tga.

the fix is easy: make the bow particle a projectile, track
it to the hero across frames (just like sword), and match
the hero's facing direction upon cast (because projectiles
can rotate cardinally however they want to).

removed ParticleBow + GameObjAnimList.
added ProjectileBow + GameObjAnimList.
updated GameObj.SetType() for ProjectileBow.
added path for ProjectileBow to Projectile.Update().
updated UseWeapon() to create a ProjectileBow,
instead of ParticleBow.

updated Battle.Damage() to prevent ProjectileBow
from attempting to deal 0 damage to the hero upon
overlap, which caused the hit sound fx, but didn't
put the hero into a hit state (lol).

playtested, the shooting arrow animation is CRISPY.