Bumpers ++
improve the bouncing algorithm.

rewrote GetOppositeDirection() to actually return the opposite
direction, including diagonals, because we can just call
GetCardinalDirection() to remove the diagonals later.

based on playtesting with hero, the best bouncing algorithm
is simply to set the bouncee's compMove.direction to the opposite
direction, without any math involved at all. this works as
expected for the hero, but not for roomObjs, so the tradeoff
right now is in favor of predictable bounces with the hero.