Fix NonBlocking Obj Interactions with Blocking RoomObjs
Flamethrower can be pushed by a belt into a blocking
roomObject, like a wall, and the flamethrower will pass
right thru the wall or blocking object. obvs, we dont
want this to happen.

the problem is a little complicated to pick apart:
interactions() happen, then movement is projected().
inside interact(): flamethrower hit belt, was pushed,
hit wall, was stopped. frame over.
then on the next frame, the flamethrower interacted() with
belt and was pushed, then interacted with wall and was stopped.
this caused the visual jittering. push, stop, etc..

made flamethrower blocking, cause that's the easiest fix.
turns out, that's all I needed to do to fix flamethrowers.

improved Key_005.
simplified Key_006.
various changes and lots of playtesting.