Fix Fairys Lifetime
we only need to set the fairy's lifetime to 0 for her to stick around
forever. currently we've implemented a seperate path for her in
gameObj.update(), using a type check, which resets her life to 100,
because REASONS? -> this is amazingly stupid and shows that i wasn't
thinking about how the system was designed while i was migrating
the fairy from a projectile to a roomObject. that's fixed now.
and of course it works just fine.