ART++
migrated sprites to The New Layout.

what is The New Layout? larger sprite sheets,
allowing for more room for sprite animations,
and the sprites are now organized according to
color palette or type, so forest world &
dungeon objects and npcs all live on the same
sheet - this will make it much easier to color
everything later on (and automate that process
with ColorSchemer, but shhh... secret stuff).

included these master ART files in the project,
in the new "ART" folder, masters and finals included.
so now I can push art commits. the first of many.
but this is not what git is really for, so.. meh

some minor cleanup codewise too.