Migrate GameObjects, MenuItems, RoomData + Sprites to The New System
created AnimationFrames.cs, a class which contains the
game's animation frames - suprise!!!
removed Functions_GameObjAnimList.cs from repo.
removed cursed castle sheet, shop sheet, main sheet.
added entities sheet, forest level sheet, uiItems sheet.
updated Assets to properly load new sheets (for Dx and UWP).
note that pets are currently broken, cause they were on
the main sheet, and that sheet doesn't exist anymore.
removed Pets from the game, temporarily, because they conflict
with the new layout and system. later on they will be re-implemented
as gameobjects that track to the hero, similar to the Fairy object.
we could even spawn this Pet Object on the hero (without collisions)
everytime he enters a room, so we are resetting the obj to a known
state upon a room build(), but that's planning for another time.
fixed Vendors and Vendor Ads.
fixed particles, projectiles, and pickups.
removed stat menu items. lol.
various comments noting bugs and issues.
migrated ALL RoomData and TestRooms to The New System.

For context:
This was the single largest commit I've made to this project.
This commit took around 10 hours total to prepare and log.
It fundamentally changes how sprite animations are handled,
and BROKE ALOT OF STUFF. Which will be fixed in future commits.
But, currently the game is in a visually corrupted state, almost
all sprites are incorrectly displayed, so the game plays kinda
like a SNES cartridge that you found on the side of the road.

This was a massive commit, with alot of consequences that I'll
be discovering for the next few weeks, if not months.