Improve Pickup Sprites
when the pickup reaches it's lowest visual position,
the shadow underneath the pickup should enlarge by 1 pixel
in diameter, to simulate the shadow growing as the object
approaches the ground.

also modified the pickup arrow sprite to bounce higher
over the shadow, making the animation clearer, which required
shortening the arrow sprites a little.

also changed how the arrow pickup sprite is visually
displayed at it's highest position - made it a brighter
color of white to grab the player's attention.

playtested, looking much better.