Editor Design ++
Based on the idea that Levels and Rooms are distinct,
Editor has been split into level and room parts.
room editor still acts as room builder did, designing
individual rooms according to specific rules which the
dungeon designer systems use to construct dungeons.
levels, however, are connected by hand and designed
by hand, and have no such system rules applied to them.

added editor mockups for level (with and without NPCs),
and for the slightly modified room editor.

the good news is that multiple functions can be shared
between the editor screens, the widgets have been simplified
and will be reused, and the objects widget (which is big
and complicated and error prone) will be split into
multiple smaller widgets that are passive programmatically,
meaning they wont process input. this vastly simplifies them,
and the input can be processed on the editor screen,
and depending on which editor screen is open the input can
be modified to suit that screen's needs (for example,
checking to see if the editor is adding a chest to non-chestable
room, and popping an error message telling them no!).

we've also created a place for dungeon entrances to live,
alongside the statues inside the dungeon that thematically
represent the boss or the dungeon - the eye statue reminding
the player of the boss, the giant floating eye, and the
dungeon entrance has a single eye, and the mob enemies have
single eyes.. this grouping helps to remind me to visually
design all these objects with this coupling in mind.