Redefine Objects Widget
created WidgetObject.cs file, which will contain all the classes
related to the new object widgets that are being developed.
once they're designed and integrated, i'll split this file apart
or maybe not depending on what the design calls for. if you
haven't noticed by now, IDGAF about subscribing blindly to paradigms.
this is just the fastest way to dev the data and systems right now.

created WidgetObjects_Dungeon, and populated it with testing objs.
created Functions_Widget, with a method for positioning all gameObjs
for any passed ObjWidget. tested, working as expected.

setup the new Dungeon Objects Widget, got it animating and drawing
in Room Editor, the first step towards matching the code to the
mockups and intended functionality.