Implement 'F6 - Hide Editor Widgets'
added HideEditorWidgets boolean to Flags class.
this boolean will control if editor widgets process
input, update, and draw methods.

updated Debug.HandleTopMenuInput() to flip above
boolean upon F6 button press. updated RoomEditor
screen to check above boolean prior to processing
any widget's update, input, or draw methods.

playtested, and working for roomEditor.

(thoughts - a pioneering new section of my commits..)
this function allows editors the ability to view the level or room as the
player would - EXACTLY - allowing the ability to fill levels to the
edge of the camera with objects, so a level looks visually different from a
dungeon in the sense that the surrounding darkness is never seen outside on
a world level. it's there, but just covered with game objs.

an addition to the above thought, is that the darkness inside a dungeon
could simply be set to different color based on the level type, so if the
level was outside, the background could switch to a grass color, and then
we wouldn't need to have a TON of basic green grass sprites all over the screen.
instead, we can sprinkle some animated flowers and grass around on the
background and it will look the same, with a significantly lower draw time.

if we couple that optimization with the floor sprites optimization inside,
(since all rooms are set to specific sizes now), we could drop draw times
dramatically for levels, possibly allowing for larger levels. hmm..