Design Cheats Screen
designed cheats screen with a few goals in mind:
1. should be simple and easy to read.
2. should be easily expandable later on.
3. should fit into the current design language,
meaning window frame, title, selection box linked
to menuItems or gameObjects, and text describing
the cheats effect.

the mockup i've designed to meet these goals is
pretty good in my opinion. it's simple, and will
be easy to implement in code too.