Fix Pit Teeth
updated level sheet with modified placement of pit teeth.
we'll just keep them as 16x16 pixel objects, instead of
turning them into 16x8, and saving a frame in the spritesheet.
the additional complications in the codebase to make the
16x8 pit teeth work just isn't worth the effort, rn imo.

fixed the animationFrames for pit teeth top and bottom.
playtested, working as expected.