Fix Entity Textures
so right now we're forcing all particles, projectiles,
and pickups to have their sprite textures set to the
entities sheet. however, we are 'themeing' some particles,
which means they are part of the dungeon sheet and match
that dungeon sheet's overall color and theme.

for example, in one dungeon the pits have bubbles, so the
visual effect is that the pit is filled with water or
lava. but in a desert dungeon, the pits could have swirling
particles, while the 'pit' sprite is sand, showing up
as quicksand, but with the same effect (pulling actors in).

the fix here is simply to pull particles and projectiles
out of the final group check in Obj.SetType(), and instead
set the texture per path in SetType().

moved Pickup group check in Pickup paths.
fixed PitBubbles (the whole reason to do this).
migrated all other particles, including the dungeon
map and key, which should also live on the dungeon sheet.
better region'd out SetType().Particles section.
migrated projectiles, including exploding barrel
projectile which lives on the dungeon sheet.

playtested, exploding barrels and pit bubbles
working properly now, along with other entities.
region'd out Obj.SetType() more for readability.